requirements.txt
flask==3.0.3
pandas>=2.2.2
numpy>=2.0.0
scikit-learn>=1.5.0
joblib>=1.4.2
watchdog==4.0.1
gunicorn==22.0.0
tensorflow-cpu==2.18.0
scapy==2.5.0